Move the class to the appropriate directory
