+++ dubbo-xds/src/main/java/org/apache/dubbo/xds/credentials/InternalServiceProviders.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.apache.dubbo.xds;
+package org.apache.dubbo.xds.credentials;
 
 import java.util.List;
