+++ dubbo-xds/src/main/java/org/apache/dubbo/xds/credentials/XdsCredentialsProvider.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.apache.dubbo.xds;
+package org.apache.dubbo.xds.credentials;
 
 import io.grpc.ChannelCredentials;
 import io.grpc.Internal;
