--- dubbo-xds/src/main/resources/META-INF/services/org.apache.dubbo.xds.XdsCredentialsProvider
@@ -1,3 +1,0 @@
-org.apache.dubbo.xds.credentialsProvider.GoogleDefaultXdsCredentialsProvider
-org.apache.dubbo.xds.credentialsProvider.InsecureXdsCredentialsProvider
-org.apache.dubbo.xds.credentialsProvider.TlsXdsCredentialsProvider
