+++ dubbo-xds/src/main/resources/META-INF/services/org.apache.dubbo.xds.credentials.XdsCredentialsProvider
@@ -1,0 +1,3 @@
+org.apache.dubbo.xds.credentials.GoogleDefaultXdsCredentialsProvider
+org.apache.dubbo.xds.credentials.InsecureXdsCredentialsProvider
+org.apache.dubbo.xds.credentials.TlsXdsCredentialsProvider
